ta: pkcs11: Fix PKCS11 TA failures in ECDSA processing
Fixes problem in ECDSA processing when ECDSA signing is chosen
it need not depend on key size but a chosen algorithm need to be
returned. Whereas during verification the key size need to be
returned. Fixing these subtelities passes the PKCS11 test on
platform using newer version of the global platform. The patch
try to be backward compatible by conditionally compiling when
the defines are expected for newer implementations.