drivers: Add stm32mp1 remoteproc driver
This driver is in charge of the configuration of registers and memories
of the remote processor.
Based on the device tree it stores information about memories assigned
to the remote processor.
Then It ensures coherency between the memory registered and the addresses
of the firmware segments to load.
it starts and stops the remote processor core.

Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
Change-Id: If2b1e8ff975c39f2d4d9569653f811ba96de74f6