drivers: regulator: register to dt_driver
Adds regulator_dt_register() for regulator drivers to register regulator
instances based on the DT description of the platform.

Regulator instances may not be created and initialized when
regulator_dt_register() returns. When a regulator depends on a
supply this latter may not yet be registered and initialized.
The framework will resolve the regulator dependencies later.

At OP-TEE core last initcall stage, a debug message informs
in case of remaining unresolved regulator dependency. Used resources
are released and no error status is returns to the system.

regulator_dt_register() uses a dedicated struct regu_dt_desc
ABI to get the description of the regulator to be registered.

Co-developed-by: Pascal Paillet <p.paillet@foss.st.com>
Signed-off-by: Pascal Paillet <p.paillet@foss.st.com>
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>