plat-stm32mp1: default disable reserved shared memory
Changes stm32mp15 default configuration to not embed OP-TEE's reserved
static shared memory. This change is aligned with a recent change in
TF-A [1] that also default disables the related configuration switch.
Note that TF-A/stm32mp1 deprecates this configuration in order to fully
remove its support in the future [2].

Note that when CFG_CORE_RESERVED_SHM is disabled the default 2MB area
is now used by OP-TEE as secure memory.

Link: https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?id=fb1d3bd9330ce70f735a344dd4223faffb261118 [1]
Link: https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit/?id=12e683a68049f6a3d0985a2cd1564e00115e809f [2]
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Gatien Chevallier <gatien.chevallier@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@foss.st.com>