Extract pkcs2tee_key_type functions to helpers
Extracts the two pkcs2tee_key_type functions from the symm and asymm
implementation, unite the implementation and place it in the
pkcs11_helpers.c file so that other files can make use of this.

Renames the function to the more appropriate pkcs2tee_object_type, as
we're returning a TEE Object Type.

Signed-off-by: Lars Feyaerts <lars@bitbiz.be>