Adds Transient Token capability
Adds the concept of a transient token which, instead of creating a TEE
Persistent Object that's stored persistently, allows for storing a TEE
Transient Object that's stored transiently (i.e. volatile).

A transient token reuses the existing database structures used by the
persistent token, but instead of writing it to persistent storage using
a TEE Persistent Object, it just keeps it in volatile memory. A
transient token doesn't use the db_objs field because it only allows for
storing one object (for now) using the pkcs11_object's key_handle member.

Whether a token is transient or persistent is determined by the
PKCS11_CKFS_HW_SLOT flag. Reasoning behind this is that persistent
tokens are backed by hardware storage (e.g. flash) while transient
tokens aren't.

Add one such token during the pkcs11 initialization. The ability to add
new transient tokens when a new object is written to the existing
transient token, is not present in this commit but might be desirable
to allow for a dynamic amount of transient tokens.

Signed-off-by: Lars Feyaerts <lars@bitbiz.be>