ci: Rust updates
Now that [1] is merged, Rust tests are enabled by default for QEMUv8 and
there is no need for a separate Rust job in CI, so remove it. On the
other hand, a couple of fixes are needed:

- Update PATH so that the cargo command (which is installed locally
during the build of the Rust SDK) can be found.

- Disable Rust in the BTI+MTE+PAC test because the Rust examples fail to
build with the supplied toolchain:

 /usr/local/bin/../lib/gcc/aarch64-unknown-linux-uclibc/12.2.0/../../../../aarch64-unknown-linux-uclibc/bin/ld.bfd: /tmp/rustcmQty55/libcompiler_builtins-76fca0633b54e12b.rlib(45c91108d938afe8-cpu_model.o): in function `init_have_lse_atomics':
 /cargo/registry/src/index.crates.io-6f17d22bba15001f/compiler_builtins-0.1.101/./lib/builtins/cpu_model.c:1075: undefined reference to `getauxval'
 ...

- Also disable Rust in the test that enables ftrace, because the
signature_verification-rs command just hangs in this configuration.

Link: https://github.com/OP-TEE/build/pull/717 [1]
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>