core: tests: add a notification test watchdog
Add test watchdog for asynchronous notifications where a timer interrupt
triggers an asynchronous notification in the normal world kernel driver.
The normal world kernel driver responds by doing a
OPTEE_MSG_CMD_DO_BOTTOM_HALF call for bottom half processing. The
watchdog checks that there has been a response for each timer interrupt but
doesn't take any measures if a response is missing.

The purpose of the test is to exercise asynchronous notifications.
Feedback is limited to debug prints on the UART so eventual regressions
will not get caught by xtest unless there is a crash.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>