core: add is_nexus() and refactor is_unpaged()
Add the function is_nexus() to tell if an address is an address
available to the nexus when non-secure virtualization is enabled
(CFG_NS_VIRTUALIZATION=y). The function is stubbed to return true for
non-null arguments if non-secure virtualization isn't enabled, else
false.

Make the argument for is_unpaged() const void * for consistency with the
is_nexus() function. The stubbed version of is_unpaged() when paging
isn't enabled is updated to perform a NULL check on the argument.

Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>