mk/lib.mk: add library to link line only when it does contain objects
This commit addresses a build issue when the output directory is not
cleaned from a previous build. Here is a test case:

 # Start from a clean state
 $ rm -rf ./out
 # (1) Build for Arm, with unwinding enabled
 $ make -s -j$(nproc) ARCH=arm PLATFORM=vexpress-qemu_armv8a O=out CFG_UNWIND=y && echo OK
 OK
 # (2) Build for RISC-V, with unwinding enabled too
 $ make -s -j$(nproc) ARCH=riscv PLATFORM=virt O=out CFG_UNWIND=y && echo OK
 OK
 # (3) Build for Arm again but with unwinding disabled
 $ make -s -j$(nproc) ARCH=arm PLATFORM=vexpress-qemu_armv8a O=out CFG_UNWIND=n
 aarch64-linux-gnu-ld.bfd: skipping incompatible out/ldelf-lib/libunw/libunw.a when searching for -lunw
 aarch64-linux-gnu-ld.bfd: cannot find -lunw: No such file or directory
 make: *** [ldelf/link.mk:60: out/ldelf/ldelf.elf] Error 1
 make: *** Waiting for unfinished jobs....

In step (3), the libunw.a file leftover from step (2) causes a problem
because it is not generated again for the current ARCH (due to
CFG_UNWIND=n, so there is effectively nothing to build). Yet it is
unconditionally added to the link line by mk/lib.mk although
CFG_UNWIND=n. Therefore change the logic in mk/lib.mk to deal with that.

Step (2) causes no error because due to CFG_UNWIND=y and due to
dependencies (different source files, different cross compiler), the
archive file is re-created.

Note that it is not OK to simply guard the inclusion of mk/lib.mk with
CFG_UNWIND in core/core.mk and ldelf/ldelf.mk because we still want
the library headers to be accessible (no conditionals on #include <...>).

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Etienne Carriere <etienne.carriere@foss.st.com>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>