core: rpmb: add platform specific hook to avoid rpmb key exchange
Currently the core already check if the key was written and don't
trigger a new tee_rpmb_write_and_verify_key(). But this does not prevent
us from attacks which involves: re-soldering the eMMC to re-trigger the
key exchange.

This new platform specific function can prevent such attacks, e.g. by
checking an e-fuse and return true if the exchange was already done.

Signed-off-by: Marco Felsch <m.felsch@pengutronix.de>