driver: crypto: hisilicon: fix timeout error
In last commit, I changed the code from using cnt to using timeout to
determine when to break the circulate.However, this do not take account
of time cost in function(such as qm_mb_read()), only time cost in delay.
This will cause timeout when running the code.So, I rolled back this
change.

Signed-off-by: yuzexi <yuzexi@hisilicon.com>