core: arm64: increase STACK_ABT_SIZE from 1024 to 3072 when log level is 0
When adding "make check CFG_WITH_PAGER=y CFG_TEE_CORE_LOG_LEVEL=0" to
the QEMUv8 CI job, I noticed that OP-TEE fails to boot and hangs with no
message printed on the console. The root cause is memory corruption of
the translation tables triggered by a stack overflow. Indeed, the pager
uses the abort stack to handle unmapped pages, and therefore it requires
quite a bit of stack space. The log level is not very relevant.
Therefore, fix the issue by removing the particular case for log level 0.

More debugging info:

 build$ make -j$(nproc) CFG_WITH_PAGER=y CFG_TEE_CORE_LOG_LEVEL=0 \
                        CFG_CORE_ASLR=n
 build$ aarch64-linux-gnu-nm -n ../optee_os/out/arm/core/tee.elf
 ...
 000000000e115000 B __nozi_start
 000000000e115000 b thread_user_kdata_page
 000000000e116000 b xlat_tables_ul1
 000000000e118000 b xlat_tables
 000000000e11d000 b base_xlation_table
 000000000e11d100 B __nozi_end
 000000000e11d100 B __nozi_stack_start
 000000000e11d100 b stack_abt
 000000000e11e200 B stack_tmp
 ...
 build$ make run-only
 optee_qemuv8$ gdb-multiarch
 (gdb) symbol-file optee_os/out/arm/core/tee.elf
 (gdb) target remote localhost:1234
 (gdb) p sizeof(base_xlation_table)
 $1 = 256
 (gdb) watch *(char [256]*)base_xlation_table
 (gdb) c  # 5 times
 Thread 1 hit Hardware watchpoint 1: *(char [256]*)base_xlation_table
 (gdb) bt

At this point the call stack is:

 hash_sha256_check()
 fobj_load_page()
 pager_deploy_page()
 pager_get_page()
 tee_pager_handle_fault()
 abort_handler()
 el1_sync_abort()

This code is indeed not supposed to touch base_xlation_table, it does
so due to the overflow of stack_abt.

Suggested-by: Jens Wikander <jens.wiklander@linaro.org>
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Joakim Bech <joakim.bech@linaro.org>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>